+++ polyamorous/lib/polyamorous/activerecord_6.0_ruby_2/reflection.rb
@@ -1,2 +1,2 @@
 # active_record_6.0_ruby_2/reflection.rb
-require 'polyamorous/activerecord_5.2.0_ruby_2/reflection'
+require 'polyamorous/activerecord_5.2_ruby_2/reflection'