+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Drop support for Active Record 5.0, 5.1, and 5.2.0.
+  PR [#1073](https://github.com/activerecord-hackery/ransack/pull/1073)
+  
 * Drop support for rubies under 2.3.
   PR [#1070](https://github.com/activerecord-hackery/ransack/pull/1070)